backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
pyserial>=3.5
python-dotenv>=1.0.0
websockets>=12.0
